Move local mutations into entity stores
