+++ src/entities/card/model/store.spec.tsx
@@ -1,10 +1,28 @@
 import { renderHook } from "@testing-library/react";
-import { beforeEach, describe, expect, it } from "vitest";
+import { beforeEach, describe, expect, it, vi } from "vitest";
 import { createJSONStorage, type StateStorage } from "zustand/middleware";
 
 import { createCard } from "@/test/factories";
 import { useCard, useCards, useCardsByDeckId } from "./hooks";
-import { cardStore, clearRemoteCards, replaceRemoteCards } from "./store";
+import {
+  cardStore,
+  clearRemoteCards,
+  createLocalCard,
+  deleteLocalCard,
+  deleteLocalCardsByDeckId,
+  editLocalCard,
+  replaceRemoteCards,
+} from "./store";
+
+const cardInput = (id: string, deckId = "deck") => ({
+  id,
+  deckId,
+  uid: "uid",
+  frontText: "front",
+  backText: "back",
+  tags: [],
+  uniqueKey: `key-${id}`,
+});
 
 const createMemoryStorage = (initial: Record<string, string> = {}): StateStorage => {
   const values = new Map(Object.entries(initial));
@@ -25,6 +43,7 @@ describe("Card store", () => {
   beforeEach(() => {
     useMemoryStorage();
     cardStore.setState({ remoteCards: [], localCards: [] });
+    vi.useRealTimers();
   });
 
   it("replaces and clears only the remote Card collection", () => {
@@ -91,5 +110,27 @@ describe("Card store", () => {
     await cardStore.persist.rehydrate();
 
     expect(cardStore.getState().localCards).toEqual([]);
+  });
+
+  it("creates, edits, and deletes a local Card synchronously", () => {
+    vi.spyOn(Date, "now").mockReturnValueOnce(10).mockReturnValueOnce(20);
+    const createdCard = createLocalCard(cardInput("local"));
+
+    expect(createdCard).toEqual(expect.objectContaining({ id: "local", createdAt: 10, updatedAt: 10 }));
+
+    const updatedCard = editLocalCard({ id: "local", uid: "uid", frontText: "updated" });
+    expect(updatedCard).toEqual(expect.objectContaining({ frontText: "updated", createdAt: 10, updatedAt: 20 }));
+
+    deleteLocalCard("local");
+    expect(cardStore.getState().localCards).toEqual([]);
+  });
+
+  it("deletes local Cards by Deck", () => {
+    createLocalCard(cardInput("first", "deck-a"));
+    createLocalCard(cardInput("second", "deck-b"));
+
+    deleteLocalCardsByDeckId("deck-a");
+
+    expect(cardStore.getState().localCards.map(({ id }) => id)).toEqual(["second"]);
   });
 });
