+++ src/entities/deck/index.ts
@@ -1,9 +1,9 @@
 export { createDeck, deleteDeck, editDeck, fetchDecks, subscribeDecks } from "./api/firestore";
-/** @public */
-export { createLocalDeck, deleteLocalDeck, editLocalDeck } from "./api/local";
 export { generateDeckId } from "./model/id";
 export { CATEGORY, getCategory, isHighlightLanguage } from "./model/rules";
 export { useDeck, useDecks } from "./model/hooks";
 export { clearRemoteDecks } from "./model/store";
+/** @public */
+export { createLocalDeck, deleteLocalDeck, editLocalDeck } from "./model/store";
 /** @public */
 export type { Deck, DeckCreateInput, DeckEdit, DeckId, LocalDeckCreateInput } from "./model/types";
